refactor(plugin-openclaw): extract delegate session-namespace resolution
delegate-runtime.ts crossed the 1200-line cap for new source files. The
namespace helpers are a coherent unit — recall, observe, and the lifecycle
flushes all answer the same 'which namespace is this session bound to?'
question from a host event, a durable binding history, and the daemon default —
so they move to delegate-namespaces.ts. Keeping them together is what stops the
three hook paths from drifting apart, and from drifting away from the
memory-slot search that resolves the same way.
